requirements.txt
smtplib
email
python-dotenv
pycoingecko
rich
cx-Freeze
